+++ src/main/java/com/notenoughmail/kubejs_tfc/util/helpers/ducks/IOpenSimplex2dMixin.java
@@ -1,9 +1,0 @@
-package com.notenoughmail.kubejs_tfc.util.helpers.ducks;
-
-import dev.latvian.mods.rhino.util.HideFromJS;
-
-@HideFromJS
-public interface IOpenSimplex2dMixin {
-
-    void kubejs_tfc$SetSeed(long seed);
-}